+++ firmware/communicator/stm32/cmake/arm-none-eabi-gcc.cmake
@@ -2,6 +2,10 @@ set(CMAKE_SYSTEM_NAME Generic)
 set(CMAKE_SYSTEM_PROCESSOR arm)
 set(CMAKE_TRY_COMPILE_TARGET_TYPE STATIC_LIBRARY)
 
+if(DEFINED ENV{ARM_GNU_TOOLCHAIN_ROOT})
+    list(PREPEND CMAKE_PROGRAM_PATH "$ENV{ARM_GNU_TOOLCHAIN_ROOT}/bin")
+endif()
+
 find_program(CMAKE_C_COMPILER arm-none-eabi-gcc REQUIRED)
 find_program(CMAKE_ASM_COMPILER arm-none-eabi-gcc REQUIRED)
 find_program(CMAKE_OBJCOPY arm-none-eabi-objcopy REQUIRED)